requirements.txt
firebase-admin==6.5.0
python-dotenv==1.0.1
typer==0.12.5
rich==13.8.1
tabulate==0.9.0
pytest==8.3.2
